+++ web/modules/custom/ai_screening_project_track/src/Computer/WebformSubmissionProjectTrackToolComputer.php
@@ -3,6 +3,8 @@
 namespace Drupal\ai_screening_project_track\Computer;
 
 use Drupal\Core\Entity\EntityInterface;
+use Drupal\ai_screening_project_track\Exception\InvalidValueException;
+use Drupal\ai_screening_project_track\Helper\FormHelper;
 use Drupal\ai_screening_project_track\ProjectTrackToolComputerInterface;
 use Drupal\ai_screening_project_track\ProjectTrackToolInterface;
 use Drupal\ai_screening_project_track\Status;
@@ -25,6 +27,29 @@ final class WebformSubmissionProjectTrackToolComputer implements ProjectTrackToo
    */
   public function compute(ProjectTrackToolInterface $tool, EntityInterface $entity): void {
     assert($entity instanceof WebformSubmissionInterface);
+
+    $calculated = [];
+
+    // Loop over form fields values and calculate sums.
+    foreach ($entity->getData() as $value) {
+      try {
+        if (!empty($value)) {
+          $fieldValues = FormHelper::getIntegers($value);
+          foreach (array_keys($calculated + $fieldValues) as $key) {
+            $calculated[$key] = ($calculated[$key] ?? 0) + ($fieldValues[$key] ?? 0);
+          }
+        }
+
+      }
+      catch (InvalidValueException $e) {
+      }
+    }
+
+    // Add the sums to track data.
+    $toolData = $tool->getToolData();
+    $toolData['summed_dimensions'] = $calculated;
+    $tool->setToolData($toolData);
+
     $tool->setProjectTrackToolStatus(Status::IN_PROGRESS);
   }