+++ web/modules/custom/ai_screening_project_track/src/Helper/ProjectTrackHelper.php
@@ -11,9 +11,12 @@ use Drupal\ai_screening_project_track\Evaluation;
 use Drupal\ai_screening_project_track\Event\ProjectTrackToolComputedEvent;
 use Drupal\ai_screening_project_track\ProjectTrackInterface;
 use Drupal\ai_screening_project_track\ProjectTrackStorageInterface;
+use Drupal\ai_screening_project_track\ProjectTrackToolStorageInterface;
 use Drupal\ai_screening_project_track\Status;
 use Drupal\core_event_dispatcher\Event\Theme\ThemeEvent;
 use Drupal\core_event_dispatcher\ThemeHookEvents;
+use Drupal\taxonomy\TermStorageInterface;
+use Drupal\webform\WebformSubmissionStorageInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
@@ -30,13 +33,39 @@ final class ProjectTrackHelper extends AbstractHelper implements EventSubscriber
    */
   private readonly ProjectTrackStorageInterface|EntityStorageInterface $projectTrackStorage;
 
+  /**
+   * The project track storage.
+   *
+   * @var \Drupal\ai_screening_project_track\ProjectTrackToolStorageInterface|\Drupal\Core\Entity\EntityStorageInterface
+   */
+  private readonly ProjectTrackToolStorageInterface|EntityStorageInterface $projectTrackToolStorage;
+
+
+  /**
+   * The term storage.
+   *
+   * @var \Drupal\taxonomy\TermStorageInterface|\Drupal\Core\Entity\EntityStorageInterface
+   */
+  private readonly TermStorageInterface|EntityStorageInterface $termStorage;
+
+  /**
+   * The webform submission storage.
+   *
+   * @var \Drupal\Core\Entity\EntityStorageInterface|\Drupal\webform\WebformSubmissionStorageInterface
+   */
+  private readonly WebformSubmissionStorageInterface|EntityStorageInterface $submissionStorage;
+
   public function __construct(
     private readonly ProjectTrackToolHelper $projectTrackToolHelper,
+    private readonly ProjectTrackTypeHelper $projectTrackTypeHelper,
     EntityTypeManagerInterface $entityTypeManager,
     LoggerChannel $logger,
   ) {
     parent::__construct($logger);
     $this->projectTrackStorage = $entityTypeManager->getStorage('project_track');
+    $this->projectTrackToolStorage = $entityTypeManager->getStorage('project_track_tool');
+    $this->termStorage = $entityTypeManager->getStorage('taxonomy_term');
+    $this->submissionStorage = $entityTypeManager->getStorage('webform_submission');
   }
 
   /**
@@ -44,7 +73,7 @@ final class ProjectTrackHelper extends AbstractHelper implements EventSubscriber
    *
    * @param \Drupal\ai_screening_project_track\ProjectTrackInterface $track
    *   *   The track.
-   * @param string $key
+   * @param string|null $key
    *   The key.
    *
    * @return mixed
@@ -55,20 +84,22 @@ final class ProjectTrackHelper extends AbstractHelper implements EventSubscriber
    *
    * @see self::hasTrackData()
    */
-  public function getToolData(ProjectTrackInterface $track, ?string $key = NULL): mixed {
-    try {
-      $data = $track->getToolData();
+  public function getToolsData(ProjectTrackInterface $track, ?string $key = NULL): mixed {
+    $data = [];
+    $projectTrackToolIds = $this->projectTrackToolStorage->getQuery()
+      ->accessCheck(FALSE)
+      ->condition('project_track_id', $track->id())
+      ->sort('delta')
+      ->execute();
 
-      return NULL === $key ? $data : ($data[$key] ?? NULL);
-    }
-    catch (\Exception $exception) {
-      $this->logException($exception, __METHOD__, [
-        'track' => $track,
-        'key' => $key,
-      ]);
+    $projectTrackTools = $this->projectTrackToolStorage->loadMultiple($projectTrackToolIds);
+
+    /** @var \Drupal\ai_screening_project_track\Entity\ProjectTrackTool $projectTrackTool */
+    foreach ($projectTrackTools as $projectTrackTool) {
+      $data[$projectTrackTool->id()] = $projectTrackTool->getToolData();
     }
 
-    return NULL;
+    return $data;
   }
 
   /**
@@ -83,55 +114,6 @@ final class ProjectTrackHelper extends AbstractHelper implements EventSubscriber
    */
   public function getEvaluationOptions(): array {
     return Evaluation::asOptions();
-  }
-
-  /**
-   * Check if track has data.
-   *
-   * @param \Drupal\ai_screening_project_track\ProjectTrackInterface $track
-   *   The track.
-   * @param string $key
-   *   The key.
-   *
-   * @return bool
-   *   True if data for key exists.
-   */
-  public function hasTrackData(ProjectTrackInterface $track, string $key): bool {
-    try {
-      $data = $track->getToolData();
-
-      return array_key_exists($key, $data);
-    }
-    catch (\Exception $exception) {
-      $this->logException($exception, __METHOD__, [
-        'track' => $track,
-        'key' => $key,
-      ]);
-    }
-
-    return FALSE;
-  }
-
-  /**
-   * Add data to track.
-   *
-   * @param \Drupal\ai_screening_project_track\ProjectTrackInterface $track
-   *   The track.
-   * @param string $key
-   *   The key.
-   * @param mixed $value
-   *   The value.
-   */
-  public function setTrackData(ProjectTrackInterface $track, string $key, mixed $value): void {
-    try {
-      $track->setToolData([$key => $value] + $track->getToolData());
-    }
-    catch (\Exception $exception) {
-      $this->logException($exception, __METHOD__, [
-        'track' => $track,
-        'key' => $key,
-      ]);
-    }
   }
 
   /**
@@ -156,7 +138,7 @@ final class ProjectTrackHelper extends AbstractHelper implements EventSubscriber
    *
    * @throws \Drupal\Core\Entity\EntityStorageException
    */
-  public function deleteProjectTracks(array $projectTracks) {
+  public function deleteProjectTracks(array $projectTracks): void {
     foreach ($projectTracks as $projectTrack) {
       $this->projectTrackToolHelper->deleteTools($projectTrack);
 
@@ -166,13 +148,63 @@ final class ProjectTrackHelper extends AbstractHelper implements EventSubscriber
 
   /**
    * Event handler.
+   *
+   * @throws \Drupal\Core\Entity\EntityStorageException
    */
   public function projectTrackToolComputed(ProjectTrackToolComputedEvent $event): void {
-    // @todo Do something with the tool's track?
-    /*
     $tool = $event->getTool();
     $track = $tool->getProjectTrack();
-     */
+    $tools = $this->getToolsData($track);
+    $summedDimensions = [];
+
+    // Sum up all project track tools.
+    /** @var \Drupal\ai_screening_project_track\Entity\ProjectTrackTool $tool */
+    foreach ($tools as $tool) {
+      // Sum up each dimension.
+      foreach (array_keys($summedDimensions + $tool['summed_dimensions']) as $key) {
+        $summedDimensions[$key]['sum'] = ($summedDimensions[$key] ?? 0) + ($tool['summed_dimensions'][$key] ?? 0);
+        $summedDimensions[$key]['undecidedThreshold'] = $this->projectTrackTypeHelper->getThreshold($track->getType()->id(), $key, Evaluation::UNDECIDED);
+        $summedDimensions[$key]['approvedThreshold'] = $this->projectTrackTypeHelper->getThreshold($track->getType()->id(), $key, Evaluation::APPROVED);
+      }
+    }
+
+    // Determine the Evaluation as a sum of all dimensions and track thresholds.
+    // There is probably some smarter way I just couldn't find one.
+    $result = [];
+
+    // Loop over all dimensions and determine which threshold the sum value
+    // matches.
+    foreach ($summedDimensions as $summedDimension) {
+      if ($summedDimension['sum'] < $summedDimension['undecidedThreshold']) {
+        $result['refuse'] = TRUE;
+      }
+      if (($summedDimension['undecidedThreshold'] > $summedDimension['sum']) &&
+          ($summedDimension['sum'] < $summedDimension['approvedThreshold'])) {
+        $result['undecided'] = TRUE;
+      }
+      if ($summedDimension['sum'] > $summedDimension['approvedThreshold']) {
+        $result['approved'] = TRUE;
+      }
+    }
+
+    // After loop we set the evaluation only ending on approved if other keys
+    // were not found.
+    if (array_key_exists('refuse', $result)) {
+      $evaluation = Evaluation::REFUSED;
+    }
+    elseif (array_key_exists('undecided', $result)) {
+      $evaluation = Evaluation::UNDECIDED;
+    }
+    elseif (array_key_exists('approved', $result)) {
+      $evaluation = Evaluation::APPROVED;
+    }
+
+    $trackConfig = $track->getConfiguration();
+    $trackConfig['sums'] = $summedDimensions;
+    $trackConfig['evaluation'] = $evaluation ?? [];
+
+    $track->setConfiguration($trackConfig);
+    $track->save();
   }
 
   /**