+++ web/modules/custom/ai_screening_project_track/src/Helper/ProjectTrackTypeHelper.php
@@ -13,6 +13,7 @@ use Drupal\ai_screening_project_track\Evaluation;
 use Drupal\ai_screening_project_track\Exception\InvalidConfigurationException;
 use Drupal\core_event_dispatcher\Event\Form\FormAlterEvent;
 use Drupal\core_event_dispatcher\FormHookEvents;
+use Drupal\taxonomy\Entity\Term;
 use Drupal\taxonomy\TermForm;
 use Drupal\taxonomy\TermInterface;
 use Drupal\taxonomy\TermStorageInterface;
@@ -29,6 +30,7 @@ final class ProjectTrackTypeHelper implements EventSubscriberInterface {
   const string CONFIGURATION_KEY_DIMENSIONS = 'dimensions';
   public const string BUNDLE_TERM_PROJECT_TRACK = 'project_track_type';
   private const string THRESHOLD_KEY_SEPARATOR = '-';
+
   /**
    * The term storage.
    *
@@ -169,6 +171,41 @@ final class ProjectTrackTypeHelper implements EventSubscriberInterface {
         sprintf('Configuration value "%s" must be a list', self::CONFIGURATION_KEY_DIMENSIONS)
       );
     }
+  }
+
+  /**
+   * Determine the max possible values that a project track can achieve.
+   *
+   * @throws \Drupal\ai_screening_project_track\Exception\InvalidValueException
+   */
+  public function getProjectTrackTypeMaxPossible(Term $projectTrackType): array {
+    $currentMax = [];
+    $currentMaxTrack = [];
+    $webforms = $projectTrackType->get('field_webform')->referencedEntities();
+    // Loop over each tool in the project track.
+    foreach ($webforms as $webform) {
+      $elements = $webform->getElementsDecodedAndFlattened();
+      // Look at each field in the webform.
+      foreach ($elements as $element) {
+        // Currently we only look at weighted radios.
+        if ('ai_screening_weighted_radios' === $element['#type'] && array_key_exists('#options', $element)) {
+          // Get all options for the field.
+          foreach ($element['#options'] as $optionValues => $option) {
+            $values = FormHelper::getIntegers($optionValues);
+            // Get possible max for each element.
+            foreach ($values as $key => $value) {
+              $currentMax[$key] = max($currentMax[$key] ?? 0, $value);
+            }
+
+          }
+        }
+        $currentMaxTrack = array_map(function () {
+          return array_sum(func_get_args());
+        }, $currentMaxTrack, $currentMax);
+      }
+    }
+
+    return $currentMax;
   }
 
   /**