+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+- [PR-67](https://github.com/itk-dev/ai-screening/pull/67)
+  - Add max values calculation.
+  - Add track tool sum calculation
+  - Add track calculation
 - [PR-66](https://github.com/itk-dev/ai-screening/pull/66)
   Add threshold configuration page and apply it to graphs.
 - [PR-65](https://github.com/itk-dev/ai-screening/pull/65)